frontend/requirements.txt
fastapi
uvicorn
langchain-huggingface==0.0.3
langchain-community==0.0.36
langchain-core==0.1.52
langchain-groq==0.1.3
pypdf==4.2.0
supabase-client==2.5.0